realized the logger wasnt initialized before I was attempting to use it
